move serialization test so that internal state can be checked
