+++ src/test/java/com/fasterxml/jackson/databind/TestJDKSerialization.java
@@ -193,31 +193,13 @@ public class TestJDKSerialization extends BaseMapTest
         assertEquals(JavaType.class, t.getRawClass());
     }
 
-    public void testLRUMap() throws Exception
-    {
-        LRUMap<String,Integer> map = new LRUMap<String,Integer>(32, 32);
-        map.put("a", 1);
-        assertEquals(1, map.size());
-
-        byte[] bytes = jdkSerialize(map);
-        LRUMap<String,Integer> result = jdkDeserialize(bytes);
-        // transient implementation, will be read as empty
-        assertNull(result.get("a"));
-        assertEquals(0, result.size());
-
-        // but should be possible to re-populate
-        assertNull(result.put("a", 2));
-        assertEquals(Integer.valueOf(2), result.get("a"));
-        assertEquals(1, result.size());
-    }
-
     /*
     /**********************************************************
     /* Helper methods
     /**********************************************************
      */
     
-    protected byte[] jdkSerialize(Object o) throws IOException
+    public static byte[] jdkSerialize(Object o) throws IOException
     {
         ByteArrayOutputStream bytes = new ByteArrayOutputStream(2000);
         ObjectOutputStream obOut = new ObjectOutputStream(bytes);
@@ -227,7 +209,7 @@ public class TestJDKSerialization extends BaseMapTest
     }
 
     @SuppressWarnings("unchecked")
-    protected <T> T jdkDeserialize(byte[] raw) throws IOException
+    public static <T> T jdkDeserialize(byte[] raw) throws IOException
     {
         ObjectInputStream objIn = new ObjectInputStream(new ByteArrayInputStream(raw));
         try {
